chore: add github actions workflow for deployment and pull request diffs (#90)
This adds Github actions workflows for

- Deployment to `staging` from `main`
- Diffs for pull requests against the `staging` env.

It's using Github Deployment environments to setup the secrets
https://github.com/winglang/wing.cloud/settings/environments

First attempt was to use the Wing Github Action, but it turned out to be
too unflexible for this case here. So, I went with a similar custom
setup. The plan is to iterate on the Github Action with the learnings
from this setup once this is fully baked.

For end2end tests, we can setup a similar setup for a dedicated AWS
testing account. But that's another pull request and will probably
require a bit more thinking.

Once this is merged, it should automatically deploy a staging
environment from `main` (potentially needs some babysitting for the
rollout).

## Tasks

- [x] Currently, it's using the same AWS permissions as for diffs and
deployments, which should be changed to read only for diffs.
- [x] We need a dedicated dnsimple automation user for wing.cloud which
only has access to relevant domains