feat: expand cloudfront logs (#173)
This PR does a few things:

- It's eventually fixing the cloudfront logs deployment issues around
S3. This was due to the S3 ACLs sometimes not being applied before
cloudfront was starting to be deployed. The explicit `dependable` fixes
this
- Adds an example with test (helped to isolate the issue) with a loaded
Athena table ready to query
- Adds an Athena table to actually being able to query the logs if we
need to

This could (and maybe should) be part of a custom platform as well. It
could automatically enrich existing distributions within an app and
inject the additional needed resources. However, right now it's
apparently not possible to use wing modules for custom platforms. We
should consider this once it's implemented.
https://github.com/winglang/wing/issues/4937