feat: parameter resource (#145)
started to prototype a resource for using SSM as configuration source.
This could potentially resolve the cyclic errors
https://github.com/winglang/wing/issues/4906 issue. Could be useful
beyond that for other configuration values we have.

Would be nice if there was a way to get this type safe (as in, know
within the editor which values have been configured).

related issues:

- https://github.com/winglang/wing/issues/2726
- https://github.com/winglang/wing.cloud/issues/125
- https://github.com/winglang/wing/issues/1668
- https://github.com/winglang/wing/issues/2583

besides from the functionality, I added a folder structure
`components/<name>` as a proposal. I think this could be useful for
structuring the code base a bit more into modules / components /
services / whatever you want to call it, with the goal to:

- make it easier to navigate / grasp
- have all pieces colocated 
- including examples / tests for straightforward testing in isolation
(local and tf-aws)
- encourage boundaries between these components (and potentially enforce
them as well)

Structures like this worked reasonably well for me so far, both in AWS
CDK and cdktf, so I think this could beneficial within wing as well.

Not exactly sure what the candidates would be, but probably these, maybe
more 🤷

- api
- proxy
- github
- users
- fly

thoughts?