feat: store and retrieve secrets (#143)
Saving customer secrets (which are used by `cloud.Secret`) and
retrieving them from the DB.
On AWS, every secret requests a new encryption key from KMS, which is
used to encrypt the secret value and is stored in the DB together with
the encrypted data. When retrieving the secret, we decrypt that
encryption key through KMS and use that decrypted key to decrypt the
actual secret.