feat: allow only one app for a repository
resolves: https://github.com/winglang/wing.cloud/issues/171

Wing Cloud app's name is the repository name.
Each repository can have only one Wing Cloud app associated with it.
In order to enforce it, preventing the user from creating two applications with the same name (same repository name).

In the future, we will probably allow multiple apps for the same repository, in this case we will add a counting index appended to the repo (app) name if one already exists.