fix: staging deploy
copy and paste issue. `base_ref` is empty on main, only in pr context
available.

github.base_ref	string	The base_ref or target branch of the pull request in a workflow run. This property is only available when the event that triggers a workflow run is either pull_request or pull_request_target.