feat: endpoints support (#242)
support endpoints in wing.cloud:
* environment image collects the endpoints from the console server
(fixed right now to `cloud.Api/cloud.Website/ex.ReactApp`
* adds a reverse proxy in the environment image which routes requests
based on the `Host` header
* the reverse proxy also terminates SSL with a certificate found on the
FS
* environment image reports back the list of endpoints to wing.cloud
* for each endpoint we create a DNS record `endpoint-id.wingcloud.io` ->
`fly-app.fly.dev`
* the certificate (from DNSimple) is created in advance for now since it
takes time and it cant be deleted only expired

for context why the dns solution was chosen (copied from
[(here):](https://github.com/winglang/wing.cloud/issues/191#issuecomment-1827773971)

> Initially i started with the proxied solution since i figured it would
be easier, but then i got blocked with supporting websocket based
endpoints. basically the proxy solution would mean to have an API
gateway with a lambda backend which would receive a request and forward
it to fly.io. however its not clear to me how would that work with
websockets since the serverless AWS solution for websockets is to remove
the actual concept of sockets and let the user implement
connect/disconnect/onmessage methods. in that case i'm not sure how to
have an additional stateful websocket connection which is forwarded to
fly.io (and can communicate in both directions)