fix: create app when main.w entrypoint is not in the root folder
When user creates an app we are now fetching all *.main.w files and searching for main.w in all folder.

If main.w found anywhere in the app repository, it is being set as the app entrypoint. Otherwise, an empty string is assigned and the deployment will show an error.

resolves: https://github.com/winglang/wing.cloud/issues/321