chore: single branch deployment (#363)
this changes the deployment for both `production` and `staging` from the
`main` branch. `staging` is deployed first, once it's successful the
`production` deployment will start.

the pull request diffs are made both against `staging` and `production`.
we only really care about `production`, however `staging` needs to be
successfully deployed as well. If it gets too annoying, we can still
turn the `staging` one off.

the `aws tests` on `main` (right now doing a terraform create / destroy
only) is running individually and is not prepened be before the deploy.
reasoning there is, that the main value there is, that we ensure that
the application can be created from scratch and fully destroyed again.