chore: rebuild environments with out of date sha (#411)
When an environment build is done, we now check if the last commit of
that branch is the one deployed in the environment. If its not, we
rebuild the environment with the new commit.

Part of #202. Still missing a FIFO queue. New commits are already
immediately starting a new flow without waiting for current ones to
finish.