chore: add data store for user apps
The motivation for this change was that newly added apps weren't immediately shown when the user navigated back to the apps list page.

The reason was that although the app.list query was invalidated, the refetch happened only when the user was in this page.

having the apps list data at the top level of the application and not tied to a specific component allows us to reload the data from anywhere in our application.