feat: add attribution to signed up users (#650)
This PR closes the loop of our first version of the attribution
mechanism.
Whenever we direct anyone to our assets (Github, website) we use a short
link that directs them to wing.cloud with a utm_source param.
We use Plausible in wing.cloud to log this utm_source and then
wing.cloud redirects the user to the end destination (Github, our
website, etc.).

When a user eventually signs up to the wing console we want then to pass
through wing.cloud again so Plausible can recognize it is the same
person and know that they signed up and attribute this sign up to the
utm_source that originally brought the user to us.

This PR is the missing piece of the system - it adds the part that makes
the user pass though wing.cloud after the signup so it can attribute the
signup to the utm_source. wing.cloud then redirects the user the
localhost console.

The currect attribution system has a few shortcomings that will need to
be addressed in the future:
1. Since we have both winglang.io and wing.cloud we cannot see the
entire user journey in Plausible accross both websites. This will be
remedied when we move to a single domain.
2. Since we don't connect the user's email or any other identifier to
the utm_source, we cannot know which sign ups from a source (or even how
many) are retained users and not just users who signed up. To fix this
we need to either report this data to Plausible somehow or track it
ourselves in our DB using our own cookies.